Reduce some common boiler-plate
