+++ functions/test/rate-limiter/redis-rate-limiter.test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const Redis = require('ioredis');
+const { getToday } = require('../../rate-limiter/util');
 
 jest.mock('ioredis');
 
@@ -52,13 +53,6 @@ describe('RedisRateLimiter', () => {
     jest.restoreAllMocks();
   });
 
-  function getToday() {
-    const today = new Date();
-    const dd = String(today.getDate()).padStart(2, '0');
-    const mm = String(today.getMonth() + 1).padStart(2, '0');
-    const yyyy = today.getFullYear();
-    return yyyy + mm + dd;
-  }
 
   describe('Basic functionality', () => {
     test('should initialize with zero counts', async () => {
